Add RakeOE::PrjFileCache #self.join_regex_keys_for!(a_string)
It's now possible to extend all variables inside a prj.rake file with a
regular expression that contains a pattern for the applicable platform.
The contents of these variables will be added to the contents of the base
variable if the chosen platform matches the regular expression.

With this functionality it's now easy to e.g. compile different
source folders for different platforms, use different CFLAGS/LDFLAGS, export
different include files, even use different project type, etc, etc.

Signed-off-by: Daniel Schnell <dschnell@posteo.de>