Use object files of application instead of the app-lib when building the normal executable. The app lib is still used in case of unit testing.
This is necessary if one builds projects with e.g. functions for irq
handlers in the application project which are weakly defined in other
parts of the system and are supposed to be overwritten via a user defined
function. These will not be defined if the linker option -gc-sections is
used together with compilation options -fdata-sections
-ffunction-sections resulting of missing functions in the binary.

Signed-off-by: Daniel Schnell <dschnell@posteo.de>